Add option to set indicator brightness when disconnected
If you set an LED to be lit when an indicator state is off, then it
would previously always be lit when the keyboard was not connected to
any host, which was not ideal. To fix this, this adds a new property to
set a different LED brightness when disconnected, which defaults to 0.